+++ src/main/java/nextstep/subway/common/ExceptionAdvice.java
@@ -1,6 +1,6 @@
-package nextstep.subway.line.ui.exception;
+package nextstep.subway.common;
 
-import nextstep.subway.line.ui.LineController;
+import nextstep.subway.line.application.exception.LineNotFoundException;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
@@ -8,26 +8,23 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-@RestControllerAdvice(basePackageClasses = LineController.class)
-public class LineExceptionAdvice {
-
-    private static final String LINE_ALREADY_REGISTERED = "이미 등록되어 있는 노선입니다.";
-    private static final String LINE_NOT_EXISTS = "존재하지 않는 노선입니다.";
+@RestControllerAdvice
+public class ExceptionAdvice {
 
     @ExceptionHandler(DataIntegrityViolationException.class)
     protected ResponseEntity handleDataIntegrityViolationException(
         DataIntegrityViolationException e) {
-        return ResponseEntity.status(HttpStatus.CONFLICT).body(LINE_ALREADY_REGISTERED);
+        return ResponseEntity.status(HttpStatus.CONFLICT).build();
     }
 
     @ExceptionHandler(LineNotFoundException.class)
     protected ResponseEntity handleLineNotFoundException(LineNotFoundException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(e.getMessage());
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 
     @ExceptionHandler(EmptyResultDataAccessException.class)
     protected ResponseEntity handleEmptyResultDataAccessException(
         EmptyResultDataAccessException e) {
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(LINE_NOT_EXISTS);
+        return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
     }
 }
